Factor out DropDelegate into own file
